LayoutTests/media/track/video-track-add-visible-expected.txt


RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplay)
RUN(track = video.addTextTrack("subtitles"))
RUN(track.mode = "showing")
RUN(cue = new VTTCue(0, video.duration, "cue text"))
RUN(track.addCue(cue))
EXPECTED (internals.shadowRoot(video).textContent.includes("cue text") == 'true') OK
END OF TEST
